+++ is_valid/is_cond.py
@@ -1,5 +1,5 @@
 from .base import Predicate
-from .is_eq import to_pred
+from .to_pred import to_pred
 from .is_fixed import is_fixed