+++ is_valid/is_eq.py
@@ -27,19 +27,3 @@ class is_eq(Predicate):
             if data == context(self._value) else
             (self._not_valid_exp if explain else False)
         )
-
-
-def to_pred(value):
-    if isinstance(value, Predicate):
-        return value
-    elif isinstance(value, dict):
-        from .is_dict_where import is_dict_where
-        return is_dict_where(value)
-    elif isinstance(value, list):
-        from .is_list_where import is_list_where
-        return is_list_where(*value)
-    elif isinstance(value, tuple):
-        from .is_tuple_where import is_tuple_where
-        return is_tuple_where(*value)
-    else:
-        return is_eq(value)