+++ is_valid/is_optional.py
@@ -1,6 +1,6 @@
 from .base import Predicate
 from .is_none import is_none
-from .is_eq import to_pred
+from .to_pred import to_pred
 
 
 class is_optional(Predicate):